--- src/ceasiompy/smtrain/func/parameter.py
@@ -1,10 +1,0 @@
-# Imports
-
-from pydantic import BaseModel
-
-
-# Classes
-class Parameter(BaseModel):
-    name: str
-    min_value: float
-    max_value: float
