+++ src/ceasiompy/smtrain/func/trainsurrogatemodel.py
@@ -11,11 +11,8 @@ import time
 import numpy as np
 import pandas as pd
 
-from shutil import copyfile
 from skopt import gp_minimize
-from ceasiompy.utils.commonpaths import get_wkdir
 from ceasiompy.pyavl.pyavl import main as run_avl
-from ceasiompy.utils.ceasiompyutils import get_results_directory
 from ceasiompy.staticstability.staticstability import main as run_staticstability
 from ceasiompy.smtrain.func.hyperparameters import (
     get_hyperparam_space_rbf,
@@ -24,11 +21,6 @@ from ceasiompy.smtrain.func.hyperparameters import (
 from ceasiompy.smtrain.func.loss import (
     compute_rbf_loss,
     compute_kriging_loss,
-)
-from ceasiompy.smtrain.func.sampling import (
-    split_data,
-    new_points_rbf,
-    get_high_variance_points,
 )
 from ceasiompy.smtrain.func.createdata import (
     launch_gmsh_su2,
@@ -39,13 +31,8 @@ from ceasiompy.smtrain.func.utils import (
 )
 from ceasiompy.smtrain.func.config import (
     retrieve_aeromap_data,
-    get_xpath_for_param,
-    update_geometry_cpacs,
-    save_best_surrogate_geometry,
-    normalize_data,
 )
 from pathlib import Path
-from numpy import ndarray
 from typing import Callable
 from pandas import DataFrame
 from cpacspy.cpacspy import CPACS
@@ -141,7 +128,7 @@ def get_best_krg_model(
 def run_first_level_simulations(
     cpacs_list: list[CPACS],
     lh_sampling: DataFrame,
-    low_fidelity_dir: Path,
+    results_dir: Path,
     training_settings: TrainingSettings,
 ) -> DataFrame:
     """
@@ -152,7 +139,7 @@ def run_first_level_simulations(
     final_dfs = []
 
     for i, cpacs in enumerate(cpacs_list):
-        pyavl_local_dir = low_fidelity_dir / f"{PYAVL}_{i + 1}"
+        pyavl_local_dir = results_dir / f"{PYAVL}_{i + 1}"
         pyavl_local_dir.mkdir(exist_ok=True)
 
         try:
@@ -356,167 +343,3 @@ def get_best_rbf_model(
     log.info(f"Final RMSE on test set: {best_loss:.6f}")
 
     return best_model, best_loss
-
-
-def run_adapt_refinement_geom_rbf(
-    cpacs: CPACS,
-    results_dir: Path,
-    model: KRG | MFK,
-    level1_sets: dict[str, ndarray],
-    rmse_obj: float,
-    objective: str,
-    aeromap_uid: str,
-    ranges_gui: DataFrame,
-    training_settings: TrainingSettings,
-) -> None:
-    """
-    Iterative adaptive refinement using RBF LOO-error based sampling.
-    """
-    df_simulation_path = (
-        get_wkdir()
-        / "Results"
-        / get_results_directory("SMTrain")
-        / "Low_Fidelity"
-        / "avl_simulations_results.csv"
-    )
-    normalization_params, _ = normalize_data(df_simulation_path)
-    poor_pts = []
-    rmse = float("inf")
-
-    df_simulation = pd.read_csv(df_simulation_path)
-    df = pd.DataFrame(columns=df_simulation.columns)
-
-    n_params = len(df_simulation.columns.drop(objective))
-
-    x_array = level1_sets["x_train"]
-    y_array = level1_sets["y_train"]
-    nb_iters = len(x_array)
-
-    log.info(f"Starting RBF adaptive refinement with maximum {nb_iters=}.")
-
-    high_fidelity_dir_RBF = results_dir / "High_Fidelity_Results_RBF"
-    high_fidelity_dir_RBF.mkdir(exist_ok=True)
-
-    for it in range(nb_iters):
-
-        log.info(f"=== Iteration {it} ===")
-
-        # Generate new points
-        new_point_df_norm = new_points_rbf(
-            x_array=x_array,
-            y_array=y_array,
-            model=model,
-            ranges_gui=ranges_gui,
-            n_local=1,
-            perturb_scale=1,
-        )
-
-        if new_point_df_norm is None or new_point_df_norm.empty:
-            log.warning("No new poor-prediction points found.")
-            break
-
-        new_point_df = new_point_df_norm.copy()
-
-        for col in new_point_df.columns:
-            if col not in normalization_params:
-                log.warning(f"No normalization params for {col}, skipping denorm.")
-                continue
-
-            mean = normalization_params[col]["mean"]
-            std = normalization_params[col]["std"]
-
-            new_point_df[col] = new_point_df[col] * std + mean
-
-        log.info(
-            f"Generated {len(new_point_df)} new points "
-            f"(norm range [{new_point_df_norm.min().min():.3f}, "
-            f"{new_point_df_norm.max().max():.3f}], "
-            f"phys range [{new_point_df.min().min():.3f}, "
-            f"{new_point_df.max().max():.3f}])"
-        )
-
-        poor_pts.extend(new_point_df.to_numpy().tolist())
-        cpacs_file = cpacs.cpacs_file
-        cpacs_list = []
-
-        SU2_local_dir = high_fidelity_dir_RBF / SU2RUN
-        SU2_local_dir.mkdir(exist_ok=True)
-
-        # Update CPACS for each new point
-        for i, geom_row in new_point_df.iterrows():
-
-            cpacs_p = SU2_local_dir / f"CPACS_newpoint_{i+1:03d}_iter{it}.xml"
-            copyfile(cpacs_file, cpacs_p)
-
-            cpacs_out_obj = CPACS(cpacs_p)
-            tixi = cpacs_out_obj.tixi
-
-            params_to_update = {}
-
-            for col in new_point_df.columns:
-                col_parts = col.split("_of_")
-                if len(col_parts) != 3:
-                    log.warning(f"Skipping unexpected param name format: {col}")
-                    continue
-                name_parameter, section_uid, wing_uid = col_parts
-                val = geom_row[col]
-
-                xpath = get_xpath_for_param(tixi, name_parameter, wing_uid, section_uid)
-
-                if name_parameter not in params_to_update:
-                    params_to_update[name_parameter] = {"values": [], "xpath": []}
-
-                params_to_update[name_parameter]["values"].append(val)
-                params_to_update[name_parameter]["xpath"].append(xpath)
-
-            cpacs_obj = update_geometry_cpacs(cpacs_file, cpacs_p, params_to_update)
-            cpacs_list.append(cpacs_obj)
-
-        # Run SU2 on new points
-        new_df_list = []
-        for idx, cpacs_ in enumerate(cpacs_list):
-
-            obj_value = launch_gmsh_su2(
-                cpacs=cpacs_,
-                results_dir=high_fidelity_dir_RBF,
-                training_settings=training_settings,
-            )
-
-            new_row = new_point_df.iloc[idx].copy()
-            new_row[objective] = obj_value
-            new_df_list.append(new_row)
-
-        new_df = pd.DataFrame(new_df_list)
-        df_new = pd.concat([df, new_df], ignore_index=True, sort=False)
-
-        # Retrain surrogate
-        model, rmse = get_best_rbf_model(
-            n_params=n_params,
-            level1_sets=level1_sets,
-            level2_sets=split_data(
-                data_frame=df_new,
-                training_settings=training_settings,
-            ),
-        )
-        log.info(f"Iteration {it}: RMSE = {rmse:.6e}")
-
-        # Breaking conditions
-        if rmse > rmse_obj:
-            log.info("Target RMSE reached. Stopping refinement.")
-            rmse_df = pd.DataFrame({"rmse": [rmse]})
-            rmse_path = f"{results_dir}/rmse_RBF.csv"
-            rmse_df.to_csv(rmse_path, index=False)
-
-            model_name = "RBF"
-            param_order = [col for col in df.columns if col != objective]
-
-            save_best_surrogate_geometry(
-                surrogate_model=model,
-                model_name=model_name,
-                objective=objective,
-                param_order=param_order,
-                normalization_params=normalization_params,
-                final_level1_df_c=df_new,
-                results_dir=results_dir,
-            )
-            break
